+++ core/node/eth_sender/src/eth_tx_aggregator.rs
@@ -1,3 +1,5 @@
+use std::collections::HashMap;
+
 use tokio::sync::watch;
 use zksync_config::configs::eth_sender::SenderConfig;
 use zksync_contracts::BaseSystemContractsHashes;
@@ -15,14 +17,15 @@ use zksync_l1_contract_interface::{
 use zksync_shared_metrics::BlockL1Stage;
 use zksync_types::{
     aggregated_operations::AggregatedActionType,
-    commitment::{L1BatchCommitmentMode, L1BatchWithMetadata, SerializeCommitment},
+    commitment::{L1BatchWithMetadata, SerializeCommitment},
     eth_sender::{EthTx, EthTxBlobSidecar, EthTxBlobSidecarV1, SidecarBlobV1},
     ethabi::{Function, Token},
     l2_to_l1_log::UserL2ToL1Log,
     protocol_version::{L1VerifierConfig, PACKED_SEMVER_MINOR_MASK},
     pubdata_da::PubdataSendingMode,
+    server_notification::GatewayMigrationState,
     settlement::SettlementLayer,
-    web3::{contract::Error as Web3ContractError, BlockNumber, CallRequest},
+    web3::{contract::Error as Web3ContractError, CallRequest},
     Address, L2ChainId, ProtocolVersionId, SLChainId, H256, U256,
 };
 
@@ -35,12 +38,6 @@ use crate::{
     zksync_functions::ZkSyncFunctions,
     Aggregator, EthSenderError,
 };
-
-#[derive(Debug, PartialEq)]
-pub enum GatewayMigrationState {
-    InProgress,
-    NotInProgress,
-}
 
 /// Data queried from L1 using multicall contract.
 #[derive(Debug)]
@@ -71,19 +68,17 @@ pub struct EthTxAggregator {
     pub(super) state_transition_chain_contract: Address,
     state_transition_manager_address: Address,
     functions: ZkSyncFunctions,
-    base_nonce: u64,
-    base_nonce_custom_commit_sender: Option<u64>,
     rollup_chain_id: L2ChainId,
     /// If set to `Some` node is operating in the 4844 mode with two operator
     /// addresses at play: the main one and the custom address for sending commit
-    /// transactions. The `Some` then contains the address of this custom operator
-    /// address.
-    custom_commit_sender_addr: Option<Address>,
+    /// transactions. The `Some` then contains client for this custom operator address.
+    eth_client_blobs: Option<Box<dyn BoundEthInterface>>,
     pool: ConnectionPool<Core>,
     sl_chain_id: SLChainId,
     health_updater: HealthUpdater,
     priority_tree_start_index: Option<usize>,
     settlement_layer: SettlementLayer,
+    initial_pending_nonces: HashMap<Address, u64>,
 }
 
 struct TxData {
@@ -100,29 +95,26 @@ impl EthTxAggregator {
         config: SenderConfig,
         aggregator: Aggregator,
         eth_client: Box<dyn BoundEthInterface>,
+        eth_client_blobs: Option<Box<dyn BoundEthInterface>>,
         config_timelock_contract_address: Address,
         state_transition_manager_address: Address,
         l1_multicall3_address: Address,
         state_transition_chain_contract: Address,
         rollup_chain_id: L2ChainId,
-        custom_commit_sender_addr: Option<Address>,
         settlement_layer: SettlementLayer,
     ) -> Self {
         let eth_client = eth_client.for_component("eth_tx_aggregator");
-        let functions = ZkSyncFunctions::default();
-        let base_nonce = eth_client.pending_nonce().await.unwrap().as_u64();
+        let eth_client_blobs = eth_client_blobs.map(|c| c.for_component("eth_tx_aggregator"));
 
-        let base_nonce_custom_commit_sender = match custom_commit_sender_addr {
-            Some(addr) => Some(
-                (*eth_client)
-                    .as_ref()
-                    .nonce_at_for_account(addr, BlockNumber::Pending)
-                    .await
-                    .unwrap()
-                    .as_u64(),
-            ),
-            None => None,
-        };
+        let functions = ZkSyncFunctions::default();
+
+        let mut initial_pending_nonces = HashMap::new();
+        for client in eth_client_blobs.iter().chain(std::iter::once(&eth_client)) {
+            let address = client.sender_account();
+            let nonce = client.pending_nonce().await.unwrap().as_u64();
+
+            initial_pending_nonces.insert(address, nonce);
+        }
 
         let sl_chain_id = (*eth_client).as_ref().fetch_chain_id().await.unwrap();
 
@@ -135,15 +127,14 @@ impl EthTxAggregator {
             l1_multicall3_address,
             state_transition_chain_contract,
             functions,
-            base_nonce,
-            base_nonce_custom_commit_sender,
             rollup_chain_id,
-            custom_commit_sender_addr,
+            eth_client_blobs,
             pool,
             sl_chain_id,
             health_updater: ReactiveHealthCheck::new("eth_tx_aggregator").1,
             priority_tree_start_index: None,
             settlement_layer,
+            initial_pending_nonces,
         }
     }
 
@@ -788,27 +779,37 @@ impl EthTxAggregator {
         // We may be using a custom sender for commit transactions, so use this
         // var whatever it actually is: a `None` for single-addr operator or `Some`
         // for multi-addr operator in 4844 mode.
-        let sender_addr = match (op_type, is_gateway) {
-            (AggregatedActionType::Commit, false) => self.custom_commit_sender_addr,
-            (_, _) => None,
+        let (sender_addr, is_non_blob_sender) = match (op_type, is_gateway) {
+            (AggregatedActionType::Commit, false) => self
+                .eth_client_blobs
+                .as_ref()
+                .map(|c| (c.sender_account(), false))
+                .unwrap_or_else(|| (self.eth_client.sender_account(), true)),
+            (_, _) => (self.eth_client.sender_account(), true),
         };
-        let nonce = self.get_next_nonce(&mut transaction, sender_addr).await?;
+        let nonce = self
+            .get_next_nonce(&mut transaction, sender_addr, is_non_blob_sender)
+            .await?;
         let encoded_aggregated_op =
             self.encode_aggregated_op(aggregated_op, chain_protocol_version_id);
         let l1_batch_number_range = aggregated_op.l1_batch_range();
 
-        let eth_tx_predicted_gas = match (op_type, is_gateway, self.aggregator.mode()) {
-            (AggregatedActionType::Execute, false, _) => Some(
+        let eth_tx_predicted_gas = match op_type {
+            AggregatedActionType::Execute => {
                 L1GasCriterion::total_execute_gas_amount(
                     &mut transaction,
                     l1_batch_number_range.clone(),
+                    is_gateway,
                 )
-                .await,
+                .await
+            }
+            AggregatedActionType::PublishProofOnchain => {
+                L1GasCriterion::total_proof_gas_amount(is_gateway)
+            }
+            AggregatedActionType::Commit => L1GasCriterion::total_commit_validium_gas_amount(
+                l1_batch_number_range.clone(),
+                is_gateway,
             ),
-            (AggregatedActionType::Commit, false, L1BatchCommitmentMode::Validium) => Some(
-                L1GasCriterion::total_validium_commit_gas_amount(l1_batch_number_range.clone()),
-            ),
-            _ => None,
         };
 
         let mut eth_tx = transaction
@@ -818,8 +819,8 @@ impl EthTxAggregator {
                 encoded_aggregated_op.calldata,
                 op_type,
                 timelock_contract_address,
-                eth_tx_predicted_gas,
-                sender_addr,
+                Some(eth_tx_predicted_gas),
+                Some(sender_addr),
                 encoded_aggregated_op.sidecar,
                 is_gateway,
             )
@@ -844,23 +845,19 @@ impl EthTxAggregator {
     async fn get_next_nonce(
         &self,
         storage: &mut Connection<'_, Core>,
-        from_addr: Option<Address>,
+        from_addr: Address,
+        is_non_blob_sender: bool,
     ) -> Result<u64, EthSenderError> {
         let is_gateway = self.settlement_layer.is_gateway();
         let db_nonce = storage
             .eth_sender_dal()
-            .get_next_nonce(from_addr, is_gateway)
+            .get_next_nonce(from_addr, is_non_blob_sender, is_gateway)
             .await
             .unwrap()
             .unwrap_or(0);
         // Between server starts we can execute some txs using operator account or remove some txs from the database
         // At the start we have to consider this fact and get the max nonce.
-        let l1_nonce = if from_addr.is_none() {
-            self.base_nonce
-        } else {
-            self.base_nonce_custom_commit_sender
-                .expect("custom base nonce is expected to be initialized; qed")
-        };
+        let l1_nonce = self.initial_pending_nonces[&from_addr];
         tracing::info!(
             "Next nonce from db: {}, nonce from L1: {} for address: {:?}",
             db_nonce,
@@ -876,43 +873,13 @@ impl EthTxAggregator {
     }
 
     async fn gateway_status(&self, storage: &mut Connection<'_, Core>) -> GatewayMigrationState {
-        let to_gateway = self
-            .functions
-            .server_notifier_contract
-            .event("MigrateToGateway")
-            .unwrap()
-            .signature();
-        let from_gateway = self
-            .functions
-            .server_notifier_contract
-            .event("MigrateFromGateway")
-            .unwrap()
-            .signature();
-
         let notification = storage
             .server_notifications_dal()
-            .get_last_notification_by_topics(&[to_gateway, from_gateway])
+            .get_latest_gateway_migration_notification()
             .await
             .unwrap();
 
-        notification
-            .map(|a| match self.settlement_layer {
-                SettlementLayer::L1(_) => {
-                    if a.main_topic == to_gateway {
-                        GatewayMigrationState::InProgress
-                    } else {
-                        GatewayMigrationState::NotInProgress
-                    }
-                }
-                SettlementLayer::Gateway(_) => {
-                    if a.main_topic == from_gateway {
-                        GatewayMigrationState::InProgress
-                    } else {
-                        GatewayMigrationState::NotInProgress
-                    }
-                }
-            })
-            .unwrap_or(GatewayMigrationState::NotInProgress)
+        GatewayMigrationState::from_sl_and_notification(self.settlement_layer, notification)
     }
 }