+++ core/node/eth_sender/src/tests.rs
@@ -22,7 +22,7 @@ use zksync_types::{
     helpers::unix_timestamp_ms,
     settlement::SettlementLayer,
     web3::{self, contract::Error},
-    Address, K256PrivateKey, L2ChainId, ProtocolVersionId, SLChainId, H256, U256,
+    Address, K256PrivateKey, L1BatchNumber, L2ChainId, ProtocolVersionId, SLChainId, H256, U256,
 };
 use zksync_web3_decl::client::MockClient;
 
@@ -266,6 +266,7 @@ async fn resend_each_block(commitment_mode: L1BatchCommitmentMode) -> anyhow::Re
             .eth_sender_dal()
             .get_inflight_txs(
                 tester.manager.operator_address(OperatorType::NonBlob),
+                false,
                 false
             )
             .await
@@ -322,6 +323,7 @@ async fn resend_each_block(commitment_mode: L1BatchCommitmentMode) -> anyhow::Re
             .eth_sender_dal()
             .get_inflight_txs(
                 tester.manager.operator_address(OperatorType::NonBlob),
+                false,
                 false
             )
             .await
@@ -951,4 +953,71 @@ async fn test_signing_eip712_tx() {
     let (_tx_req, _tx_hash) =
         TransactionRequest::from_bytes(tx.raw_tx.as_ref(), L2ChainId::new(chain_id).unwrap())
             .unwrap();
+}
+
+#[test_log::test(tokio::test)]
+async fn manager_monitors_even_unsuccesfully_sent_txs() {
+    let pool = ConnectionPool::<Core>::test_pool().await;
+
+    let mut tester = EthSenderTester::new(
+        pool.clone(),
+        vec![100; 100],
+        false,
+        true,
+        L1BatchCommitmentMode::Rollup,
+        SettlementLayer::L1(10.into()),
+    )
+    .await;
+
+    let _genesis_batch = TestL1Batch::sealed(&mut tester).await;
+    let l1_batch = TestL1Batch::sealed(&mut tester).await;
+    l1_batch.save_commit_tx(&mut tester).await;
+    tester.gateway_blobs.set_return_error_on_tx_request(true);
+
+    tester.run_eth_sender_tx_manager_iteration().await;
+
+    // check that we sent something and stored it in the db.
+    tester.assert_just_sent_tx_count_equals(1).await;
+    // tx should be considered in-flight.
+    tester.assert_inflight_txs_count_equals(1).await;
+
+    let mut conn = pool.connection().await.unwrap();
+    let tx = conn
+        .eth_sender_dal()
+        .get_last_sent_successfully_eth_tx_by_batch_and_op(
+            L1BatchNumber(1),
+            AggregatedActionType::Commit,
+        )
+        .await;
+    assert!(tx.is_none());
+
+    let all_attempts = conn
+        .eth_sender_dal()
+        .get_tx_history_to_check(1)
+        .await
+        .unwrap();
+    assert_eq!(all_attempts.len(), 1);
+
+    // Mark tx as successful on SL side and run eth tx manager iteration.
+    tester.confirm_tx(all_attempts[0].tx_hash, true).await;
+
+    // Check that `sent_successfully` was reset to true.
+    let tx = conn
+        .eth_sender_dal()
+        .get_last_sent_successfully_eth_tx_by_batch_and_op(
+            L1BatchNumber(1),
+            AggregatedActionType::Commit,
+        )
+        .await
+        .unwrap();
+    assert!(tx.sent_successfully);
+
+    // Check that tx is confirmed.
+    let is_confirmed = conn
+        .eth_sender_dal()
+        .get_confirmed_tx_hash_by_eth_tx_id(tx.eth_tx_id)
+        .await
+        .unwrap()
+        .is_some();
+    assert!(is_confirmed);
 }