+++ core/node/node_framework/src/implementations/layers/eth_watch.rs
@@ -79,8 +79,8 @@ impl EthWatchLayer {
             contracts_resource
                 .ecosystem_contracts
                 .state_transition_proxy_addr,
-            contracts_resource.chain_contracts_config.chain_admin,
-            contracts_resource.ecosystem_contracts.server_notifier_addr,
+            l1_ecosystem_contracts_resource.chain_admin,
+            l1_ecosystem_contracts_resource.server_notifier_addr,
             self.eth_watch_config.confirmations_for_eth_event,
             self.chain_id,
         )